* Read BMD0 files and show the textures (if any) * Support for game BLOODBAHAMUT (.DPK files)    - Pack method by Tricky Upgrade * Create NSCR files of NCGR SubImages * Improved zoom function * Improved export function * Fixed problems in Hexadecimal editor

+++ trunk/Tinke/VisorHex.cs
@@ -18,6 +18,7 @@ namespace Tinke
         bool fileEdited;
         IByteCharConverter bcc;
         string hexFile;
+        bool allowEdit;
 
         public VisorHex(sFile file, bool edit)
         {
@@ -26,6 +27,7 @@ namespace Tinke
             saveToolStripMenuItem.Enabled = edit;
 
             this.file = file;
+            allowEdit = edit;
 
             hexFile = Path.GetDirectoryName(file.path) + Path.DirectorySeparatorChar + "hex_" + Path.GetRandomFileName();
             if (new FileInfo(file.path).Length != file.size)
@@ -35,23 +37,29 @@ namespace Tinke
                 File.WriteAllBytes(hexFile, br.ReadBytes((int)file.size));
                 br.Close();
             }
+            else if (!edit)
+                hexFile = file.path;
             else
                 File.Copy(file.path, hexFile, true);
 
-            hexBox1.ByteProvider = new DynamicFileByteProvider(hexFile, false); 
+            hexBox1.ByteProvider = new DynamicFileByteProvider(hexFile, !edit); 
             encodingCombo.SelectedIndex = 0;
          }
         private void VisorHex_FormClosed(object sender, FormClosedEventArgs e)
         {
             hexBox1.Dispose();
             ((DynamicFileByteProvider)hexBox1.ByteProvider).Dispose();
-            if (File.Exists(hexFile))
+            if (File.Exists(hexFile) && allowEdit)
                 File.Delete(hexFile);
         }
 
         private void btnSave_Click(object sender, EventArgs e)
         {
+            if (!allowEdit)
+                return;
+
             fileEdited = true;
+            ((DynamicFileByteProvider)hexBox1.ByteProvider).ApplyChanges();
             String newFilePath = Path.GetDirectoryName(hexFile) + Path.DirectorySeparatorChar + "new_" + Path.GetRandomFileName();
             File.Copy(hexFile, newFilePath, true);
             file.offset = 0x00;
@@ -111,35 +119,35 @@ namespace Tinke
             for (int i = 0; i < toolStripSearchBox.Text.Length; i += 2)
                 search.Add(Convert.ToByte(toolStripSearchBox.Text.Substring(i, 2), 16));
 
-            hexBox1.Find(search.ToArray(), 0);
+            hexBox1.Find(search.ToArray(), hexBox1.SelectionStart + hexBox1.SelectionLength);
             this.Cursor = Cursors.Default;
         }
         private void shiftjisToolStripMenuItem_Click(object sender, EventArgs e)
         {
             this.Cursor = Cursors.WaitCursor;
             byte[] search = Encoding.GetEncoding("shift_jis").GetBytes(toolStripSearchBox.Text.ToCharArray());
-            hexBox1.Find(search, 0);
+            hexBox1.Find(search, hexBox1.SelectionStart + hexBox1.SelectionLength);
             this.Cursor = Cursors.Default;
         }
         private void defaultCharsToolStripMenuItem_Click(object sender, EventArgs e)
         {
             this.Cursor = Cursors.WaitCursor;
             byte[] search = Encoding.Default.GetBytes(toolStripSearchBox.Text.ToCharArray());
-            hexBox1.Find(search, 0);
+            hexBox1.Find(search, hexBox1.SelectionStart + hexBox1.SelectionLength);
             this.Cursor = Cursors.Default;
         }
         private void unicodeToolStripMenuItem_Click(object sender, EventArgs e)
         {
             this.Cursor = Cursors.WaitCursor;
             byte[] search = Encoding.Unicode.GetBytes(toolStripSearchBox.Text.ToCharArray());
-            hexBox1.Find(search, 0);
+            hexBox1.Find(search, hexBox1.SelectionStart + hexBox1.SelectionLength);
             this.Cursor = Cursors.Default;
         }
         private void unicodeBigEndianToolStripMenuItem_Click(object sender, EventArgs e)
         {
             this.Cursor = Cursors.WaitCursor;
             byte[] search = Encoding.BigEndianUnicode.GetBytes(toolStripSearchBox.Text.ToCharArray());
-            hexBox1.Find(search, 0);
+            hexBox1.Find(search, hexBox1.SelectionStart + hexBox1.SelectionLength);
             this.Cursor = Cursors.Default;
         }
     }
